+++ R/parse-preref.R
@@ -1,9 +1,5 @@
 # Parse a preref
 parse.preref <- function(lines) {
-  # Extract srcrefs (needed for error messages)
-  srcrefs <- attr(lines, 'srcref')
-  srcrefs$lloc[1] <- srcrefs$lloc[1] + 1
-
   delimited.lines <- lines[str_detect(lines, LINE.DELIMITER)]
   trimmed.lines <- str_trim(str_replace(delimited.lines, LINE.DELIMITER, ""),
     "right")
@@ -19,22 +15,25 @@ parse.preref <- function(lines) {
   ## Compress the escaped delimeters.
   elements <- str_replace_all(elements, fixed("@@"), "@")
 
-  parsed <- parse_elements(elements[-1], srcrefs)
+  parsed <- parse_elements(elements[-1])
+
   if (elements[[1]] != "") {
-    parsed$introduction <- check_rd(str_trim(elements[[1]]))
+    check_rd(NULL, elements[[1]])
+    parsed$introduction <- str_trim(elements[[1]])
   }
+
   parsed
 }
 
 # Sequence that distinguishes roxygen comment from normal comment.
 LINE.DELIMITER <- '\\s*#+\' ?'
 
-parse_elements <- function(elements, srcref) {
+parse_elements <- function(elements) {
   pieces <- str_split_fixed(elements, "[[:space:]]+", 2)
 
   parse_element <- function(tag, value) {
     tag_parser <- preref.parsers[[tag]] %||% parse.unknown
-    tag_parser(tag, value, srcref)
+    tag_parser(tag, value)
   }
 
   Map(parse_element, pieces[, 1], pieces[, 2])
@@ -46,7 +45,6 @@ parse_elements <- function(elements, srcref) {
 #'
 #' @param key the parsing key
 #' @param rest the expression to be parsed
-#' @param srcref srcref providing location of file name and line number
 #' @keywords internal
 #' @name parsers
 NULL
@@ -54,62 +52,66 @@ NULL
 #' @details \code{parse.default}: emit value unchanged.
 #' @export
 #' @rdname parsers
-parse.default <- function(key, rest, srcref) {
-  check_rd(str_trim(rest))
+parse.default <- function(key, rest) {
+  check_rd(key, rest)
+  str_trim(rest)
 }
 
-#' @details \code{parse.default}: warns about unknown tag.
+#' @details \code{parse.default}: throws errors about unknown tag.
 #' @export
 #' @rdname parsers
-parse.unknown <- function(key, rest, srcref) {
-  roxygen_warning(key, ' is an unknown key', srcref = srcref)
-  check_rd(str_trim(rest))
+parse.unknown <- function(key, rest) {
+  stop("@", key, " is an unknown key")
 }
 
 #' @details \code{parse.value}: fail if empty
 #' @export
 #' @rdname parsers
-parse.value <- function(key, rest, srcref) {
+parse.value <- function(key, rest) {
+  check_rd(key, rest)
   if (is.null.string(rest)) {
-    roxygen_stop(key, ' requires a value', srcref = srcref)
+    stop("@", key, ' requires a value')
   }
 
-  check_rd(str_trim(rest))
+  str_trim(rest)
 }
 
 #' @details \code{parse.words}: parse values into words separated by space
 #' @export
 #' @rdname parsers
-parse.words <- function(key, rest, srcref) {
-  str_split(check_rd(str_trim(rest)), "\\s+")[[1]]
+parse.words <- function(key, rest) {
+  check_rd(key, rest)
+  str_split(str_trim(rest), "\\s+")[[1]]
 }
 
 #' @details \code{parse.description}: parse mandatory name and description
 #' @export
 #' @rdname parsers
-parse.name.description <- function(key, rest, srcref) {
+parse.name.description <- function(key, rest) {
+  check_rd(key, rest)
   pieces <- str_split_fixed(rest, "[[:space:]]+", 2)
 
   name <- pieces[, 1]
   rest <- str_trim(pieces[, 2])
 
   if (is.null.string(name)) {
-    roxygen_stop(key, ' requires a name and description', srcref = srcref)
+    stop("@", key, ' requires a name and description')
   }
 
-  list(name = check_rd(name), description = check_rd(rest))
+  list(name = name, description = rest)
 }
 
 #' @details \code{parse.name}: one and only one word
 #' @export
 #' @rdname parsers
-parse.name <- function(key, name, srcref) {
-  name <- check_rd(str_trim(name))
+parse.name <- function(key, name) {
+  check_rd(key, name)
+  name <- str_trim(name)
 
   if (is.null.string(name)) {
-    roxygen_stop(key, ' requires a name', srcref = srcref)
+    stop("@", key, ' requires a name')
   } else if (str_count(name, "\\s+") > 1) {
-    roxygen_warning(key, ' ignoring extra arguments', srcref = srcref)
+    stop("@", key, ' should only have a single argument')
   }
 
   word(name, 1)
@@ -118,13 +120,18 @@ parse.name <- function(key, name, srcref) {
 #' @details \code{parse.toggle}: turn binary element on
 #' @export
 #' @rdname parsers
-parse.toggle <- function(key, rest, srcref) {
+parse.toggle <- function(key, rest) {
   TRUE
 }
 
-check_rd <- function(x) {
-  if (!rdComplete(x)) {
-    stop("Incomplete rd in tag: ", x, call. = FALSE)
+check_rd <- function(key, text) {
+  if (rdComplete(text)) return(TRUE)
+
+  text <- str_trim(text)
+  if (!is.null(key)) {
+    stop("Mismatched braces: \"@", key, " ", text, "\"", call. = FALSE)
+  } else {
+    stop("Mismatched braces: \"", text, "\"", call. = FALSE)
   }
-  x
+
 }