+++ R/parse.R
@@ -1,7 +1,7 @@
 parse_package <- function(base_path, load_code) {
   env <- load_code(base_path)
   parsed <- lapply(r_files(base_path), parse_file, env = env)
-  
+
   unlist(parsed, recursive = FALSE)
 }
 
@@ -14,7 +14,7 @@ parse_text <- function(text) {
   env <- new.env(parent = parent.env(globalenv()))
   attr(env, "hash") <- suppressWarnings(digest(env))
   setPackageName("roxygen_devtest", env)
-  
+
   sys.source(file, envir = env)
   parse_file(file, env)
 }
@@ -22,18 +22,21 @@ parse_text <- function(text) {
 parse_file <- function(file, env, env_hash = attr(env, "hash")) {
   parsed <- parse(file = file, keep.source = TRUE)
   if (length(parsed) == 0) return()
-  
+
   refs <- getSrcref(parsed)
   comment_refs <- comments(refs)
-  
+
   extract <- function(call, ref, comment_ref) {
-    preref <- parse.preref(as.character(comment_ref))
-    if (is.null(preref)) return()
+    comment_ref2 <- list(filename = file, lloc = as.vector(comment_ref))
 
-    preref$object <- object_from_call(call, env, preref)
-    preref$srcref <- list(filename = file, lloc = as.vector(ref))
+    errors_with_srcref(comment_ref2, {
+      preref <- parse.preref(as.character(comment_ref))
+      if (is.null(preref)) return()
 
-    add_defaults(preref)
+      preref$object <- object_from_call(call, env, preref)
+      preref$srcref <- list(filename = file, lloc = as.vector(ref))
+      add_defaults(preref)
+    })
   }
 
   Map(extract, parsed, refs, comment_refs)
@@ -42,7 +45,7 @@ parse_file <- function(file, env, env_hash = attr(env, "hash")) {
 # For each src ref, find the comment block preceeding it
 comments <- function(refs) {
   srcfile <- attr(refs[[1]], "srcfile")
-  
+
   # first_line, first_byte, last_line, last_byte
   com <- vector("list", length(refs))
   for(i in seq_along(refs)) {
@@ -55,7 +58,7 @@ comments <- function(refs) {
       first_byte <- refs[[i - 1]][4] + 1
       first_line <- refs[[i - 1]][3]
     }
-    
+
     last_line <- refs[[i]][1]
     last_byte <- refs[[i]][2] - 1
     if (last_byte == 0) {
@@ -67,10 +70,10 @@ comments <- function(refs) {
         last_byte <- 1e3
       }
     }
-    
+
     lloc <- c(first_line, first_byte, last_line, last_byte)
     com[[i]] <- srcref(srcfile, lloc)
   }
-  
+
   com
 }